Fix mysql migration extending

+++ tests/Migration/M190226185952TestMysqlMigration.php
@@ -15,10 +15,12 @@ class M190226185952TestMysqlMigration extends Mysql\Migration
      */
     public function safeUp()
     {
-        $this->createTable('test_table', [
-            'enum_column_from_array' => $this->enum(['1', '2', '3', '4']),
-            'enum_column_from_enum' => $this->enum(TestEnum::class)
-        ]);
+        if ($this->getDb()->getDriverName() === 'mysql') {
+            $this->createTable('test_table', [
+                'enum_column_from_array' => $this->enum(['1', '2', '3', '4']),
+                'enum_column_from_enum' => $this->enum(TestEnum::class)
+            ]);
+        }
     }
 
     /**
@@ -26,6 +28,8 @@ class M190226185952TestMysqlMigration extends Mysql\Migration
      */
     public function safeDown()
     {
-        $this->dropTable('test_table');
+        if ($this->getDb()->getDriverName() === 'mysql') {
+            $this->dropTable('test_table');
+        }
     }
 }
